[REF+FIX] dms: Refactor code to create abstract.dms.mixin to contain fields or functions that is used in some models (in these case some fields appear in directory and field and some functions about it). Fixed problem happen when access to file or folder view if record ref is set (file or folder has been created in attachment creation in partner for example)
