+++ dms/models/directory.py
@@ -30,6 +30,7 @@ class DmsDirectory(models.Model):
         "mail.thread",
         "mail.activity.mixin",
         "mail.alias.mixin",
+        "abstract.dms.mixin",
     ]
 
     _rec_name = "complete_name"
@@ -37,8 +38,6 @@ class DmsDirectory(models.Model):
 
     _parent_store = True
     _parent_name = "parent_id"
-
-    name = fields.Char(string="Name", required=True, index=True)
 
     parent_path = fields.Char(index=True)
     is_root_directory = fields.Boolean(
@@ -58,7 +57,7 @@ class DmsDirectory(models.Model):
         readonly=False,
         copy=True,
     )
-
+    # Override acording to defined in AbstractDmsMixin
     storage_id = fields.Many2one(
         compute="_compute_storage",
         comodel_name="dms.storage",
@@ -90,25 +89,6 @@ class DmsDirectory(models.Model):
         string="Subdirectories",
         auto_join=False,
         copy=False,
-    )
-    is_hidden = fields.Boolean(
-        string="Storage is Hidden", related="storage_id.is_hidden", readonly=True
-    )
-    company_id = fields.Many2one(
-        related="storage_id.company_id",
-        comodel_name="res.company",
-        string="Company",
-        readonly=True,
-        store=True,
-        index=True,
-    )
-
-    color = fields.Integer(string="Color", default=0)
-
-    category_id = fields.Many2one(
-        comodel_name="dms.category",
-        context="{'dms_category_show_path': True}",
-        string="Category",
     )
 
     tag_ids = fields.Many2many(
@@ -291,11 +271,6 @@ class DmsDirectory(models.Model):
         string="Model",
         store=True,
     )
-    res_model = fields.Char(string="Linked attachments model")
-    res_id = fields.Integer(string="Linked attachments record ID")
-    record_ref = fields.Reference(
-        string="Record Referenced", compute="_compute_record_ref", selection=[]
-    )
     storage_id_save_type = fields.Selection(related="storage_id.save_type", store=False)
 
     @api.depends("root_storage_id", "storage_id")
@@ -320,13 +295,6 @@ class DmsDirectory(models.Model):
     def _inverse_model_id(self):
         for record in self:
             record.res_model = record.model_id.model
-
-    @api.depends("res_model", "res_id")
-    def _compute_record_ref(self):
-        for record in self:
-            record.record_ref = False
-            if record.res_model and record.res_id:
-                record.record_ref = "{},{}".format(record.res_model, record.res_id)
 
     @api.depends("name", "complete_name")
     def _compute_display_name(self):
